Single-case example data

The scan package comes with a set of fictitious and authentic single-case study data, by courtesy of the particular authors.

Beretvas2008

Format

An object of class scdf (inherits from list) of length 1.

Value

Beretvas2008

Fictitious single-case intervention study. Reference: Beretvas, S., & Chung, H. (2008). An evaluation of modified R2-change effect size indices for single-subject experimental designs. Evidence-Based Communication Assessment and Intervention, 2, 120-128.

Borckardt2014

Fictitious daily pain ratings evaluating a psychological treatment. Reference: Borckardt, J. J., & Nash, M. R. (2014). Simulation modelling analysis for small sets of single-subject data collected over time. Neuropsychological Rehabilitation, 24, 492-506.

Huitema2000

Fictitious single-case intervention study. Reference: Huitema, B. E., & McKean, J. W. (2000). Design specification issues in time-series intervention models. Educational and Psychological Measurement, 60, 38-58.

Waddell2011

Fictitious single-case intervention study. Reference: Waddell, D. E., Nassar, S. L., & Gustafson, S. A. (2011). Single-Case Design in Psychophysiological Research: Part II: Statistical Analytic Approaches. Journal of Neurotherapy, 15, 160-169.

byHeart2011

Multiple-baseline (11 cases) intervention study on flash card vocabulary learning by Juergen Wilbert.

Grosche2011

Multiple-baseline (3 cases) intervention study on a direct-instructive reading intervention. Reference: Grosche, M. (2011). Effekte einer direkt-instruktiven Foerderung der Lesegenauigkeit. Empirische Sonderpaedagogik, 3, 147-161.

Grosche2014

Multiple-baseline (3 cases x 3 materials) intervention study on a reading intervention. Reference: Grosche, M., Lueke, T., & Wilbert, J. (in prep.).

GruenkeWilbert2014

Multiple-baseline (6 cases) intervention study on story mapping. Reference: Gruenke, M., Wilbert, J., & Stegemann-Calder, K. (2013). Analyzing the effects of story mapping on the reading comprehension of children with low intellectual abilities. Learning Disabilities: A Contemporary Journal, 11, 51-64.

Huber2014

Multiple-baseline (4 cases) intervention study on behavioral compliance. Scores refer to compliant behavior in percent. Reference: Huber, C. (in prep.).

Lenz2013

Fictious example from the paper Lenz, A. S. (2013). Calculating Effect Size in Single-Case Research: A Comparison of Nonoverlap Methods. Measurement and Evaluation in Counseling and Development, 46(1), 64–73.

Leidig2018

Leidig2018_l2

SSDforR2017

Example from the R package SSDforR.

Parker2011

Example from Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42(2), 284–299. https://doi.org/10.1016/j.beth.2010.08.006

Author

Juergen Wilbert