Baseline corrected tau

Kendalls tau correlation for the dependent variable and the phase variable is calculated after correcting for a baseline trend.

corrected_tau(
  data,
  dvar,
  pvar,
  mvar,
  phases = c(1, 2),
  alpha = 0.05,
  continuity = TRUE,
  repeated = TRUE
)

corrected_tauSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

alpha
Sets the p-value at and below which a baseline correction is applied.

continuity
If TRUE applies a continuity correction for calculating p

repeated
If TRUE applies the repeated median method for caluclating slope and intercept (mblm)

...
Further arguments passed to the function.

Details

This method has been proposed by Tarlow (2016). The baseline data are checked for a singificant autocorrelation (based on Kendalls Tau). If so, a non-parameteric Theil-Sen regression is applied for the baseline data where the dependent values are regressed on the measurement time. The resulting slope information is then used to predict data of the B-phase. The dependent variable is now corrected for this baseline trend and the resudials of the Theil-Sen regression are taken for further caluculations. Finally, a tau is calculated for the dependent variable and the dichtomos phase variable. The function here provides two extensions to this procedure: The more accurate Siegel repeated median regression is applied when repeated = TRUE and a continuity correction is applied when continuity = TRUE (both are the default settings).

References

Tarlow, K. R. (2016). An Improved Rank Correlation Effect Size Statistic for Single-Case Designs: Baseline Corrected Tau. Behavior Modification, 41(4), 427–467. https://doi.org/10.1177/0145445516676750

See also

Examples

dat <- scdf(c(A = 33,25,17,25,14,13,15, B = 15,16,16,5,7,9,6,5,3,3,8,11,7))
corrected_tau(dat)

#> Baseline corrected tau
#> 
#> Auto correlation in baseline:
#> tau = -0.68; p = 0.023 
#> 
#> Baseline corrected tau:
#> tau = 0.66; p = 0.001 
#> 
#> Baseline correction applied.
#>