Estimate single-case design

This functions takes an scdf an extracts design parameters. The resulting in object can be unsed to randomly create new scdf files with the same underlying parameters. This might be usefull for monte-carlo studies and bootstrapping procedures.

estimateSC(data, dvar, pvar, mvar, s = NULL, rtt = NULL, model = "JW", ...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

s
The standard deviation depcting the between case variance of the overall performance. If more than two single-cases are included in the scdf, the variance is estimated if s is set to NULL.

rtt
The reliability of the measurements. The reliability is estimated when rtt = NULL.

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

...
Further arguments passed to the lm function.

Value

A list of parameters for each single-case. Parameters include name, length, and starting measurementtime of each phase, trend level, and slope effects for each phase, mean, standarddeviation, and reliability for each case.

Examples

estimateSC(exampleABC)

#> Warning: This function is deprecated. It might be dropped without any further notice in a future update of scan.

#> $N
#> [1] 3
#> 
#> $case.names
#> [1] "Marie"    "Rosalind" "Lise"    
#> 
#> $MT
#>    Marie Rosalind     Lise 
#>       30       30       30 
#> 
#> $B.start
#>    Marie Rosalind     Lise 
#>       11       16       21 
#> 
#> $m
#> (Intercept) (Intercept) (Intercept) 
#>    62.53333    48.26667    53.27368 
#> 
#> $s
#> [1] 7.238198
#> 
#> $level
#>   phaseB   phaseB   phaseB 
#> 4.432712 2.015761 1.981551 
#> 
#> $slope
#>     phaseC     phaseC     phaseC 
#> -0.0385162  0.3651264  2.7598291 
#> 
#> $trend
#>          mt          mt          mt 
#> -0.26458953  0.06907796 -0.01215357 
#> 
#> $rtt
#> [1] -0.9062647 -0.6248008 -1.1967506
#> 
#> $design
#> $design$Marie
#>   lengths values start stop        m      level      slope   error     true
#> 1      10      A     1   10 62.53333  8.6393504 -0.2645895 99.8721 89.40261
#> 2      10      B    11   20 62.53333  4.4327119  0.1800214 99.8721 89.40261
#> 3      10      C    21   30 62.53333 -0.0385162  0.0577743 99.8721 89.40261
#>         rtt        s
#> 1 0.4723431 7.238198
#> 2 0.4723431 7.238198
#> 3 0.4723431 7.238198
#> 
#> $design$Rosalind
#>   lengths values start stop        m     level       slope    error     true
#> 1      15      A     1   15 48.26667 6.6683259  0.06907796 85.12578 99.22479
#> 2       8      B    16   23 48.26667 2.0157607  0.01973656 85.12578 99.22479
#> 3       7      C    24   30 48.26667 0.3651264 -0.49341401 85.12578 99.22479
#>         rtt        s
#> 1 0.5382397 7.238198
#> 2 0.5382397 7.238198
#> 3 0.5382397 7.238198
#> 
#> $design$Lise
#>   lengths values start stop        m    level       slope    error     true
#> 1      20      A     1   20 53.27368 7.360075 -0.01215357 115.0911 114.7296
#> 2       7      B    21   27 53.27368 1.981551  0.27859713 115.0911 114.7296
#> 3       3      C    28   30 53.27368 2.759829 -2.20062650 115.0911 114.7296
#>         rtt        s
#> 1 0.4992135 7.238198
#> 2 0.4992135 7.238198
#> 3 0.4992135 7.238198
#> 
#> 
#> $rtt.total
#> [1] 0.5027078
#>