Single-case example data sets

The scan package comes with a set of fictitious and authentic single-case study data. These data sets can be used to practice single-case data analysis and to reproduce results from the respective publications.

Beretvas2008 — Fictitious single-case intervention study from Beretvas & Chung, 2008.

Borckardt2014 — Fictitious daily pain ratings evaluating a psychological treatment from Borckardt & Nash, 2014.

byHeart2011 — Multiple-baseline (11 cases) flash card vocabulary learning (Wilbert, unpublished).

example_A24 — Number of injuries on a German autobahn before and after implementation of a speedlimit (130km/h) (Ministerium fuer Infrastruktur und Landesplanung. Land Brandenburg).

exampleA1B1A2B2 — Fictitious A1-B1-A2-B2 example dataset.

exampleA1B1A2B2_zvt — Non-fictitious A1-B1-A2-B2 example with ZVT (intelligence measure) and D2 (concentration measure) scores.

exampleAB — Fictitious AB example dataset with three cases.

exampleAB_50 — Fictitious AB example dataset (50 cases)s.

exampleAB_50.l2 — Level-2 data for exampleAB_50.

exampleAB_add — Fictitious AB example with added covariates.

exampleAB_decreasing — Fictitious AB example with an expected decreasing effect in the intervention phase.

exampleAB_mpd — Fictitious example with different phase structures for each case.

exampleAB_score — Fictitious AB example with binomial distributed score outcome.

exampleAB_simple — Simple fictitious AB example with three cases.

exampleABAB — Fictitious ABAB reversal design example.

exampleABC — Fictitious ABC example dataset.

exampleABC_150 — Fictitious ABC example (150 cases).

exampleABC_50 — Fictitious ABC example (50 cases).

exampleABC_outlier — Fictitious ABC example with outlier.

example_atd — Fictitious AB alternating treatment design.

example_stranger - Example for screen time of Stranger Things characters.

Grosche2011 — Multiple-baseline (three cases) from a direct-instructive reading intervention (Grosche, 2011).

Grosche2014 — Multiple-baseline (3×3 materials) reading intervention (Grosche, Lueke, & Wilbert, unpublished).

GruenkeWilbert2014 — Multiple-baseline (six cases) from a story mapping reading intervention (Gruenke, Wilbert, & Stegemann-Calder, 2013).

Huber2014 — Multiple-baseline (four cases) with DBR ratings from a behavioral compliance intervention (Huber, unpublished).

Huitema2000 — Fictitious single-case intervention study (Huitema & McKean, 2000).

Lenz2013 — Fictitious example (Lenz, 2013).

Leidig2018 — Multiple-baseline good behavior game intervention (Leidig et al., 2022).

Leidig2018_l2 — Level-2 data for Leidig2018 (Leidig et al., 2022).

Parker2007 — Example dataset after Parker et al. (2007).

Parker2009 — Example dataset after Parker et al. (2009).

Parker2009b — Example dataset after Parker & Vannest (2009).

Parker2011 — Example dataset after Parker et al. (2011).

Parker2011b — Example from Parker, Vannest, & Davis (2011).

SSDforR2017 — Example from the R package SSDforR.

Tarlow2017 — Fictitious single-case intervention study (Tarlow, 2017).

Waddell2011 — Fictitious single-case intervention study (Waddell, Nassar, & Gustafson, 2011).

Usage

data(exampleAB, package = "scan")

References

Beretvas, S., & Chung, H. (2008). An evaluation of modified R2-change effect size indices for single-subject experimental designs. Evidence-Based Communication Assessment and Intervention, 2, 120–128.

Borckardt, J. J., & Nash, M. R. (2014). Simulation modelling analysis for small sets of single-subject data collected over time. Neuropsychological Rehabilitation, 24, 492–506.

Gruenke, M., Wilbert, J., & Stegemann-Calder, K. (2013). Analyzing the effects of story mapping on the reading comprehension of children with low intellectual abilities. Learning Disabilities: A Contemporary Journal, 11, 51–64.

Grosche, M. (2011). Effekte einer direkt-instruktiven Foerderung der Lesegenauigkeit. Empirische Sonderpaedagogik, 3, 147–161.

Huitema, B. E., & McKean, J. W. (2000). Design specification issues in time-series intervention models. Educational and Psychological Measurement, 60, 38–58.

Lenz, A. S. (2013). Calculating Effect Size in Single-Case Research: A Comparison of Nonoverlap Methods. Measurement and Evaluation in Counseling and Development, 46(1), 64–73.

Leidig, T., Casale, G., Wilbert, J., Hennemann, T., Volpe, R. J., Briesch, A., & Grosche, M. (2022). Individual, generalized, and moderated effects of the good behavior game on at-risk primary school students: A multilevel multiple baseline study using behavioral progress monitoring. Frontiers in Education, 7. https://www.frontiersin.org/articles/10.3389/feduc.2022.917138

Parker, R. I., Hagan-Burke, S., & Vannest, K. (2007). Percentage of All Non-Overlapping Data (PAND) An Alternative to PND. The Journal of Special Education, 40(4), 194-204.

Parker, R. I., Vannest, K. J., & Brown, L. (2009). The improvement rate difference for single-case research. Exceptional Children, 75(2), 135-150.

Parker, R. I., & Vannest, K. (2009). An improved effect size for single-case research: Nonoverlap of all pairs. Behavior Therapy, 40(4), 357-367.

Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42(2), 284–299. https://doi.org/10.1016/j.beth.2010.08.006

Parker, R. I., Vannest, K. J., & Davis, J. L. (2011). Effect Size in Single-Case Research: A Review of Nine Nonoverlap Techniques. Behavior Modification, 35(4), 303-322. https://doi.org/10.1177/0145445511399147

Tarlow, K. R. (2017). An Improved Rank Correlation Effect Size Statistic for Single-Case Designs: Baseline Corrected Tau. Behavior Modification, 41(4), 427–467. https://doi.org/10.1177/0145445516676750

Waddell, D. E., Nassar, S. L., & Gustafson, S. A. (2011). Single-Case Design in Psychophysiological Research: Part II: Statistical Analytic Approaches. Journal of Neurotherapy, 15, 160–169.

Author

Juergen Wilbert