Load single-case data from files

Use the load_scdf function to load single-case data csv, excel, or yaml files.

load_scdf(
  filename,
  data = NULL,
  sort.labels = FALSE,
  cvar = "case",
  pvar = "phase",
  dvar = "values",
  mvar = "mt",
  phase.names = NULL,
  type = NA,
  ...
)

readSC.excel(...)

readSC(...)

Arguments

filename
A character string defining the file to be loaded (e.g. "SC_Anita.csv". If left empty a dialog box for choosing will be opened.

data
A data frame. As an alternative to filname.

sort.labels
If set TRUE, the resulting list is sorted by label names (alphabetically increasing).

cvar
Sets the variable name of the "case" variable. Defaults to "case".

pvar
Sets the variable name of the "phase" variable. Defaults to "phase".

dvar
Sets the variable name of the "values" variable. Defaults to "values".

mvar
Sets the variable name of the "mt" variable. Defaults to "mt".

phase.names
A character vector with phase names. Defaults to the phase names provided in the phase variable.

type
Format of the file to be loaded. Either "csv", "xlsx", "xls", "excel", "yml" is possible. By deafult (NA) the type is extracted from the file extension.

...
Further arguments passed to the repective read function (read.table, read_excel, read_yaml command.

Value

Returns a single-case data frame. See scdf to learn about the format of these data frames.

See also

Author

Juergen Wilbert

Examples

## Read SC-data from a file named "study1.csv" in your working directory
# study1 <- load_scdf("study1.csv")

## Read SC-data from a .csv-file with semicolon as field and comma as decimal separator
# study2 <- load_scdf("study2.csv", sep = ";", dec = ",")

## writeSc and readSC
filename <- file.path(tempdir(), "test.csv")
writeSC(exampleA1B1A2B2_zvt, filename)
dat <- load_scdf(filename, cvar = "case", pvar = "part", dvar = "zvt", mvar = "day")
#> Loaded 3 cases.
res1 <- describeSC(exampleA1B1A2B2_zvt)$descriptives
res2 <- describeSC(dat)$descriptives
all.equal(res1,res2)
#> [1] TRUE