Nonoverlap of all Pairs

The nap() function calculates the nonoverlap of all pairs (NAP; Parker & Vannest, 2009). NAP summarizes the overlap between all pairs of phase A and phase B data points. If an increase of phase B scores is expected, a non-overlapping pair has a higher phase B data point. The NAP equals number of pairs showing no overlap / number of pairs where ties are counted as half non-overlaps. Because NAP can take values between 0 and 100 percent where values below 50 percent indicate an inverse effect, an nap rescaled from -100 to 100 percent where negative values indicate an inverse effect is also displayed (\(nap_{rescaled} = 2 * nap - 100\)).

Usage

nap(data, dvar, pvar, decreasing = FALSE, phases = c(1, 2))

Arguments

data

A single-case data frame. See scdf() to learn about this format.

dvar

Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar

Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

decreasing

If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

phases

A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c(1,2).

Value

nap	A data frame with NAP and additional values for each case.

N	Number of cases.

References

Parker, R. I., & Vannest, K. (2009). An improved effect size for single-case research: Nonoverlap of all pairs. Behavior Therapy, 40, 357-367.

See also

Author

Juergen Wilbert

Examples

## Calculate NAP for a study with  lower expected phase B scores
## (e.g. aggressive behavior)
gretchen <- scdf(c(A = 12, 14, 9, 10, B = 10, 6, 4, 5, 3, 4))
nap(gretchen, decreasing = TRUE)
#> Nonoverlap of All Pairs
#> 
#>       Case NAP NAP Rescaled  w    p   d   R²
#>  [case #1]  94           88 22 <.05 2.2 0.56

## Request NAP for all cases from the Grosche2011 scdf
nap(Grosche2011)
#> Nonoverlap of All Pairs
#> 
#>   Case NAP NAP Rescaled  w   p      d     R²
#>    Eva  69         38.5 24 .10  0.747 0.1223
#>  Georg  48         -4.2 62 .57 -0.072 0.0013
#>   Olaf  45        -10.4 53 .66 -0.176 0.0077