Saves data into a .csv-file

This function restructures and saves single-case data into a .csv-file.

save_scdf(data, filename = NULL, ...)

writeSC(...)

Arguments

data

A single-case data frame. See scdf to learn about this format.

filename

A character string defining the output file name (e.g. "SC_data.csv".

...

Further arguments passed to write.table.

See also

Author

Juergen Wilbert

Examples

## Save single-case data to a .csv-file
filename <- filename <- file.path(tempdir(), "test.csv")
jessica <- random_scdf(design(level = .5))
save_scdf(jessica, filename)

## Save multiple cases to a .csv-file with semicolon as field and comma as decimal separator
save_scdf(Grosche2011, filename, sep = ";", dec = ",")

## save_scdf and load_scdf
save_scdf(exampleA1B1A2B2_zvt, filename)
dat <- load_scdf(filename, cvar = "case", pvar = "part", dvar = "zvt", mvar = "day")
#> Loaded 3 cases.
res1 <- describe(exampleA1B1A2B2_zvt)$descriptives
res2 <- describe(dat)$descriptives
all.equal(res1,res2)
#> [1] TRUE