Plot single-case data

This function provides a plot of a single-case or multiple single-cases.

scplot(scdf)

set_style(object, style, ...)

add_statline(object, stat, colour = NULL, ...)

set_xlabel(object, label, colour, size)

set_ylabel(object, label, colour, size, orientation)

set_xaxis(object, limits, increment, colour, size)

set_yaxis(object, limits, colour, size)

add_title(object, label, colour, size, font, align, parse)

add_caption(object, label, colour, size, font, align, wrap, margin, parse)

add_marks(object, case, positions, colour = "red", size = 1, shape = 1)

set_phasenames(object, ..., colour, size)

set_casenames(object, ..., colour, side, size)

add_text(object, case = 1, x, y, label, colour = NULL, size = NULL, angle = 0)

add_arrow(object, case = 1, x0, y0, x1, y1, length = 0.1, colour = NULL, ...)

add_grid(object, type, width, colour)

set_background(object, colour)

add_frame(object, colour = "black")

add_box(object, colour = "black", type = "solid", width = 1)

add_labels(
  object,
  colour = "black",
  size = 0.6,
  offset = 0.4,
  position = 3,
  round = 1
)

set_line(object, colour, width, type)

set_dots(object, colour, size, shape)

add_ridge(object, colour = "grey98")

set_seperator(object, colour, width, type, extent)

Arguments

scdf
A single-case data frame.

...
Further arguments passed to the function.

label
Character string.

align
Character string. One of "left", "right", "center"

parse
If TRUE, the label is interpreted as an expression. Default = FALSE.

wrap
Number that defines the maximum characters per line before a break. IF set to FALSE, no automatic linebreak is set.

x
x position

y
y position

lines
A list defining one or multiple lines or curves to be plotted. The argument is passed as a list (e.g., list(type = "median")). Some of the procedures can be refined with an additional argument (e.g., lines = list(type = "mean", trim = 0.2) adds a 20% trimmed mean line. For multiple lines, provide a list element for each line (e.g., list( list(type = "median", col = "red"), list(type = "trend", col = "blue")). Possible lines are:

"median" Separate lines for phase A and B medians.

"mean" Separate lines for phase A and B means. By default it is 10%-trimmed. Other trims can be set, using a trim parameter (e.g., lines = list(type = "mean", trim = 0.2) draws a 20%-trimmed mean line).

"trend" Separate lines for phase A and B trends.

"trendA" OLS trend line for phase A, extrapolated throughout phase B.

"trendA_bisplit" Split middle (bi-split) trend line for phase A, extrapolated throughout phase B.

"trendA_trisplit" Tukey tri-split trend line for phase A, extrapolated throughout phase B.

"maxA/minA" Line at the level of the highest or lowest phase A score.

"medianA" Line at the phase A median score.

"meanA" Line at the phase A 10%-trimmed mean score. Apply a different trim, by using the additional argument (e.g., lines = list(type = "meanA", trim = 0.2)).

"plm" Regression lines for piecewise linear regression model.

"plm.ar" Regression lines for piecewise autoregression model. The lag is specified like this: lines = list(type = "plm.ar", ar = 2). Default lag is set to 2.

"movingMean" Draws a moving mean curve, with a specified lag: lines = list(type = "movingMean", lag = 2). Default is a lag 1 curve.

"movingMedian" Draws a moving median curve, with a specified lag: lines = list(type = "movingMedian", lag = 3). Default is a lag 1 curve.

"loreg" Draws a non-parametric local regression line. The proportion of data influencing each data point can be specified using lines = list(type = "loreg"m f = 0.66). The default is 0.5.

"lty" Use this argument to define the line type. Examples are: "solid", "dashed", "dotted".

"lwd" Use this argument to define the line's thickness, e.g., lwd = 4.

"col" Use this argument to define the line's color, e.g., col = "red".

marks
A list of parameters defining markings of certain data points.

"positions" A vector or a list of vectors indicating measurement-times to be highlighted. In case of a vector, the marked measurement-times are the same for all plotted cases. In case of a list of vectors, marks are set differently for each case. The list must have the same length as there are cases in the data file.

"col" Color of the marks.

"cex" Size of the marks.

Use for example marks = list(positions = c(1, 8, 15), col = "red", cex = 3) to make the MTs one, eight and 18 appear big and red.

Value

An scdf plot object that creates a plot when printed.