Shift values in a single-case data file

Shifting the values might be helpful in cases where the measurement time is given as a time variable (see example below).

shift(data, value, var)

shiftSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

value
Number by which to shift the values

var
Character string with the name of the target variable. Defaults to the measurement time variable.

...
Further arguments passed to the function.

Value

A scdf with shifted data

See also

Examples

### Shift the measurement time for a better estimation of the intercept
ex <- shift(example_A24, value = -1996)
plm(ex)

#> Piecewise Regression Analysis
#> 
#> Dummy model:  B&L-B 
#> 
#> Fitted a gaussian distribution.
#> F(3, 19) = 46.74; p = 0.000; R² = 0.881; Adjusted R² = 0.862
#> 

#> Warning: unable to translate '<U+0394>R²' to native encoding

#>                      B     2.5%    97.5%     SE      t     p <U+0394>R²
#> Intercept      258.714  223.364  294.065 18.036 14.344 0.000           
#> Trend year       1.857   -7.947   11.662  5.002  0.371 0.715     0.0009
#> Level phase B -148.657 -193.265 -104.050 22.759 -6.532 0.000     0.2679
#> Slope phase B   -1.726  -11.926    8.474  5.204 -0.332 0.744     0.0007
#> 
#> Autocorrelations of the residuals
#>  lag    cr
#>    1  0.10
#>    2 -0.13
#>    3 -0.10
#> 
#> Formula: injuries ~ 1 + year + phaseB + interB
#>