Tau-U for single-case data

This function calculates indices of the Tau-U family as proposed by Parker et al. (2011).

tau_u(
  data,
  dvar,
  pvar,
  ties.method = "omit",
  method = "complete",
  phases = c(1, 2)
)

tauUSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

ties.method
Defines how to handle ties. "omit" (default) excludes all ties from the calculation. "positive" counts all ties as positive comparisons, while "negative" counts them as negative comparisons.

method
"complete" (default) or "parker". The latter calculates the number of possible pairs as described in Parler et al. (2011) which might lead to tau-U values greater than 1.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

...
Further arguments passed to the function.

Value

table

A data frame containing statistics from the Tau-U family, including: Pairs, positive and negative comparisons, S, and Tau

matrix

The matrix of comparisons used for calculating the statistics.

tau_u

Tau-U value.

References

Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42, 284-299.

See also

Examples

## Calculate tau-U for the example from Parker et al. (2011)
bob <- scdf(c(2, 3, 5, 3, 4, 5, 5, 7, 6), B.start = 5)
tau_u(bob)

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#> A vs. B + Trend B - Trend A           A vs. B - Trend A 
#>                   0.5555556                   0.5000000 
#> 
#> $Case1
#>                              S      D     τ    τb     Z     p
#> A vs. B - Trend A           13 28.844 0.500 0.451 1.534 0.125
#> A vs. B + Trend B           23 30.984 0.767 0.742 2.582 0.010
#> A vs. B + Trend B - Trend A 20 33.941 0.556 0.589 2.140 0.032
#> 

## Calculate tau-U with ties counted as positive
tau_u(Grosche2011$Eva, ties.method = "positive")

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#> A vs. B + Trend B - Trend A           A vs. B - Trend A 
#>                   0.4853801                   0.3548387 
#> 
#> $Eva
#>                              S       D     τ    τb     Z     p
#> A vs. B - Trend A           33 126.107 0.355 0.262 1.409 0.159
#> A vs. B + Trend B           80 163.328 0.513 0.490 2.849 0.004
#> A vs. B + Trend B - Trend A 83 171.000 0.485 0.485 2.904 0.004
#> 

## Request tau-U for all single-cases fom the Grosche2011 data
tau_u(Grosche2011)

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#> A vs. B + Trend B - Trend A           A vs. B - Trend A 
#>                   0.2194680                   0.1666474 
#> 
#> $Eva
#>                              S       D     τ    τb     Z     p
#> A vs. B - Trend A           33 126.107 0.355 0.262 1.409 0.159
#> A vs. B + Trend B           80 163.328 0.513 0.490 2.849 0.004
#> A vs. B + Trend B - Trend A 83 171.000 0.485 0.485 2.904 0.004
#> 
#> $Georg
#>                              S       D     τ    τb     Z     p
#> A vs. B - Trend A           11 196.214 0.079 0.056 0.342 0.732
#> A vs. B + Trend B           37 264.811 0.145 0.140 0.931 0.352
#> A vs. B + Trend B - Trend A 53 275.500 0.192 0.192 1.315 0.188
#> 
#> $Olaf
#>                               S       D      τ     τb      Z     p
#> A vs. B - Trend A           -10 175.442 -0.062 -0.057 -0.336 0.737
#> A vs. B + Trend B           -14 153.493 -0.113 -0.091 -0.516 0.606
#> A vs. B + Trend B - Trend A -14 190.000 -0.074 -0.074 -0.454 0.650
#>