Descriptive statistics for single-case data

The describe function provides common descriptive statistics for single-case data.

describe(data, dvar, pvar, mvar)

describeSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

...
Further arguments passed to the function.

Value

A data frame of descriptive statistics (for each single-case), i.e.: number of observations, number of missing values, measures of central tendency, variation, and trend.

See also

Author

Juergen Wilbert

Examples

## Descriptive statistics for a study of three single-cases
describe(Grosche2011)

#> Describe Single-Case Data
#> 
#> Design:  A B 
#> 
#>         Eva Georg Olaf
#>     n.A   6     7   12
#>     n.B  13    17    8
#>   mis.A   0     0    0
#>   mis.B   0     0    0
#> 
#>          Eva Georg Olaf
#>     m.A 2.68 10.47 6.04
#>     m.B 3.43  9.80 6.00
#>    md.A 2.95  8.82 6.36
#>    md.B 3.31 10.59 5.90
#>    sd.A 0.75  4.11 1.52
#>    sd.B 1.03  2.09 1.22
#>   mad.A 0.54  4.45 1.79
#>   mad.B 1.29  2.09 1.42
#>   min.A 1.46  5.82 3.02
#>   min.B 1.86  5.60 4.53
#>   max.A 3.39 17.40 7.69
#>   max.B 4.98 12.79 8.14
#> trend.A 0.01 -0.27 0.01
#> trend.B 0.04  0.04 0.00

## Descriptives of a three phase design
describe(exampleABC)

#> Describe Single-Case Data
#> 
#> Design:  A B C 
#> 
#>         Marie Rosalind Lise
#>     n.A    10       15   20
#>     n.B    10        8    7
#>     n.C    10        7    3
#>   mis.A     0        0    0
#>   mis.B     0        0    0
#>   mis.C     0        0    0
#> 
#>         Marie Rosalind   Lise
#>     m.A 52.00    52.27  52.35
#>     m.B 72.10    73.25  73.57
#>     m.C 68.00    66.43  71.33
#>    md.A 53.50    52.00  52.00
#>    md.B 72.50    72.00  73.00
#>    md.C 69.00    68.00  76.00
#>    sd.A  8.29     8.15  10.87
#>    sd.B 11.37    13.13  10.64
#>    sd.C 12.70    10.49  21.39
#>   mad.A 11.12     7.41  10.38
#>   mad.B 10.38    10.38  16.31
#>   mad.C 17.79    11.86  20.76
#>   min.A 39.00    37.00  35.00
#>   min.B 47.00    54.00  60.00
#>   min.C 51.00    52.00  48.00
#>   max.A 63.00    65.00  74.00
#>   max.B 85.00    97.00  87.00
#>   max.C 87.00    78.00  90.00
#> trend.A -1.92     0.50  -0.09
#> trend.B -0.61     0.64   1.93
#> trend.C -0.19    -2.93 -14.00

if (FALSE) {
## Write descriptive statistics to .csv-file
study <- describe(Waddell2011)
write.csv(study$descriptives, file = "descriptives.csv")
}