Dummy function to inherit global descriptions of parameters

Dummy function to inherit global descriptions of parameters

.inheritParams(
  data,
  scdf,
  dvar,
  mvar,
  pvar,
  decreasing,
  phases,
  model,
  trend,
  level,
  slope,
  ...
)

Arguments

data
A single-case data frame. See scdf to learn about this format.

scdf
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

decreasing
If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

trend
A logical indicating if a trend parameters is included in the model.

level
A logical indicating if a level parameters is included in the model.

slope
A logical indicating if a slope parameters is included in the model.

...
Further arguments passed to the function.