Hierarchical piecewise linear model / piecewise regression

The hplm function computes a hierarchical piecewise regression model.

hplm(
  data,
  dvar,
  pvar,
  mvar,
  model = "B&L-B",
  method = "ML",
  control = list(opt = "optim"),
  random.slopes = FALSE,
  lr.test = FALSE,
  ICC = TRUE,
  trend = TRUE,
  level = TRUE,
  slope = TRUE,
  fixed = NULL,
  random = NULL,
  update.fixed = NULL,
  data.l2 = NULL,
  ...
)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

method
Method used to fit your model. Pass "REML" to maximize the restricted log-likelihood or "ML" for maximized log-likelihood. Default is "ML".

control
A list of settings for the estimation algorithm, replacing the default values passed to the function lmeControl of the nlme package.

random.slopes
If random.slopes = TRUE random slope effects of the level, trend, and treatment parameter are estimated.

lr.test
If set TRUE likelihood ratio tests are calculated comparing model with vs. without random slope parameters.

ICC
If ICC = TRUE an intraclass-correlation is estimated.

trend
A logical indicating if a trend parameters is included in the model.

level
A logical indicating if a level parameters is included in the model.

slope
A logical indicating if a slope parameters is included in the model.

fixed
Defaults to the fixed part of the standard piecewise regression model. The parameter phase followed by the phase name (e.g., phaseB) indicates the level effect of the corresponding phase. The parameter 'inter' followed by the phase name (e.g., interB) adresses the slope effect based on the method provide in the model argument (e.g., "B&L-B"). The formula can be changed for example to include further L1 or L2 variables into the regression model.

random
The random part of the model.

update.fixed
An easier way to change the fixed model part (e.g., . ~ . + newvariable).

data.l2
A dataframe providing additional variables at Level 2. The scdf File has to have names for all cases and the Level 2 dataframe has to have a column named 'cases' with the names of the cases the Level 2 variables belong to.

...
Further arguments passed to the lme function.

Value

model

List containing infromation about the applied model

N

Number of single-cases.

formla

A list containing the fixed and the random formulas of the hplm model.

hplm

Object of class lme contaning the multilevel model

model.0

Object of class lme containing the Zero Model.

ICC

List containing intraclass correlation and test parameters.

model.without

Object of class gls containing the fixed effect model.

See also

Author

Juergen Wilbert

Examples

## Compute hplm model on a MBD over fifty cases (restricted log-likelihood)
hplm(exampleAB_50, method = "REML", random.slopes = FALSE)

#> Hierarchical Piecewise Linear Regression
#> 
#> Estimation method REML 
#> Slope estimation method: B&L-B 
#> 50 Cases
#> 
#> ICC = 0.199; L = 193.8; p = 0.000
#> 
#> Fixed effects (values ~ 1 + mt + phaseB + interB)
#> 
#>                    B    SE   df      t p
#> Intercept     51.615 1.291 1281 39.982 0
#> Trend mt       0.671 0.115 1281  5.843 0
#> Level phase B 12.939 0.589 1281 21.950 0
#> Slope phase B  0.860 0.119 1281  7.240 0
#> 
#> Random effects (~1 | case)
#> 
#>           EstimateSD
#> Intercept      8.265
#> Residual       5.342

## Analyzing with additional L2 variables
hplm(Leidig2018, data.l2 = Leidig2018_l2, 
     update.fixed = .~. + gender + migration + ITRF_TOTAL*phaseB, 
     slope = FALSE, random.slopes = FALSE, lr.test = FALSE)

#> Hierarchical Piecewise Linear Regression
#> 
#> Estimation method ML 
#> Slope estimation method: B&L-B 
#> 35 Cases
#> 
#> ICC = 0.344; L = 875.4; p = 0.000
#> 
#> Fixed effects (academic_engagement ~ mt + phaseB + gender + migration + ITRF_TOTAL +     phaseB:ITRF_TOTAL)
#> 
#>                               B    SE   df      t     p
#> Intercept                 3.747 0.262 2376 14.285 0.000
#> Trend mt                  0.004 0.001 2376  6.019 0.000
#> Level phase B             0.667 0.098 2376  6.808 0.000
#> gender                   -0.020 0.301   31 -0.067 0.947
#> migration                -0.300 0.193   31 -1.556 0.130
#> ITRF_TOTAL               -0.035 0.013   31 -2.674 0.012
#> Level phase B:ITRF_TOTAL -0.001 0.005 2376 -0.279 0.780
#> 
#> Random effects (~1 | case)
#> 
#>           EstimateSD
#> Intercept      0.557
#> Residual       0.785