Multivariate Piecewise linear model / piecewise regression

This function is in an experimental status. The mplm function computes a multivariate piecewise regression model.

mplm(
  data,
  dvar,
  mvar,
  pvar,
  model = "B&L-B",
  trend = TRUE,
  level = TRUE,
  slope = TRUE,
  formula = NULL,
  update = NULL,
  na.action = na.omit,
  ...
)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

trend
A logical indicating if a trend parameters is included in the model.

level
A logical indicating if a level parameters is included in the model.

slope
A logical indicating if a slope parameters is included in the model.

formula
Defaults to the standard piecewise regression model. The parameter phase followed by the phase name (e.g., phaseB) indicates the level effect of the corresponding phase. The parameter 'inter' followed by the phase name (e.g., interB) adresses the slope effect based on the method provide in the model argument (e.g., "B&L-B"). The formula can be changed for example to include further variables into the regression model.

update
An easier way to change the regression formula (e.g., . ~ . + newvariable).

na.action
Defines how to deal with missing values

...
Further arguments passed to the lm function.

Value

model

Character string from function call (see Arguments above).

full.model

Full regression model list

See also

Author

Juergen Wilbert

Examples

##
mplm(exampleAB_add, dvar = c("wellbeing", "depression"))

#> This function is in an experimental state and only implemented for testing und debugging purposes.

#> Multivariate piecewise linear model
#> 
#> Dummy model: B&L-B 
#> 
#> Coefficients:
#>                         wellbeing depression
#> (Intercept)            48.0380952  4.0857143
#> day                     0.3785714  0.1142857
#> Level Phase Medication  3.8633333 -0.7800000
#> Slope Phase Medication -0.2754945 -0.1650549
#> 
#> Formula: y ~ 1 + day + phaseMedication + interMedication
#> 
#> Type III MANOVA Tests: Pillai test statistic
#>                        Df test stat approx F num Df den Df Pr(>F)    
#> (Intercept)             1   0.89679  152.053      2     35 <2e-16 ***
#> day                     1   0.05450    1.009      2     35 0.3750    
#> Level Phase Medication  1   0.03846    0.700      2     35 0.5034    
#> Slope Phase Medication  1   0.03911    0.712      2     35 0.4975    
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1