Multivariate Piecewise linear model / piecewise regression

The mplm function computes a multivariate piecewise regression model.

mplm(
  data,
  dvar,
  mvar,
  pvar,
  model = "B&L-B",
  trend = TRUE,
  level = TRUE,
  slope = TRUE,
  formula = NULL,
  update = NULL,
  na.action = na.omit,
  ...
)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

trend
A logical indicating if a trend parameters is included in the model.

level
A logical indicating if a level parameters is included in the model.

slope
A logical indicating if a slope parameters is included in the model.

formula
Defaults to the standard piecewise regression model. The parameter phase followed by the phase name (e.g., phaseB) indicates the level effect of the corresponding phase. The parameter 'inter' followed by the phase name (e.g., interB) adresses the slope effect based on the method provide in the model argument (e.g., "B&L-B"). The formula can be changed for example to include further variables into the regression model.

update
An easier way to change the regression formula (e.g., . ~ . + newvariable).

na.action
Defines how to deal with missing values

...
Further arguments passed to the lm function.

Value

model

Character string from function call (see Arguments above).

full.model

Full regression model list

See also

Author

Juergen Wilbert

Examples

res <- mplm(Leidig2018$`1a1`, dvar = c("academic_engagement", "disruptive_behavior"))

#> Registered S3 methods overwritten by 'car':
#>   method                          from
#>   influence.merMod                lme4
#>   cooks.distance.influence.merMod lme4
#>   dfbeta.influence.merMod         lme4
#>   dfbetas.influence.merMod        lme4

print(res)

#> Multivariate piecewise linear model
#> 
#> Dummy model: B&L-B 
#> 
#> Coefficients: 
#>               academic_engagement disruptive_behavior
#> (Intercept)             2.9863014          0.76712329
#> Trend                  -0.2157534          0.08219178
#> Level Phase B           2.1209515         -1.30956252
#> Slope Phase B           0.2186752         -0.08027001
#> 
#> Formula: y ~ 1 + mt + phaseB + interB
#> 
#> Type III MANOVA Tests: Pillai test statistic
#>               Df test stat approx F num Df den Df    Pr(>F)    
#> (Intercept)    1  0.220968  11.0621      2     78 5.897e-05 ***
#> Trend          1  0.031657   1.2750      2     78    0.2852    
#> Level Phase B  1  0.221748  11.1123      2     78 5.672e-05 ***
#> Slope Phase B  1  0.031469   1.2672      2     78    0.2874    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

## also report standardized coefficients:
print(res, std = TRUE)

#> Multivariate piecewise linear model
#> 
#> Dummy model: B&L-B 
#> 
#> Coefficients: 
#>               academic_engagement disruptive_behavior
#> (Intercept)             2.9863014          0.76712329
#> Trend                  -0.2157534          0.08219178
#> Level Phase B           2.1209515         -1.30956252
#> Slope Phase B           0.2186752         -0.08027001
#> 
#> Standardized coefficients: 
#>               academic_engagement disruptive_behavior
#> (Intercept)              0.000000           0.0000000
#> Trend                   -5.979091           5.1071682
#> Level Phase B            0.521845          -0.7224559
#> Slope Phase B            5.973756          -4.9167203
#> 
#> Formula: y ~ 1 + mt + phaseB + interB
#> 
#> Type III MANOVA Tests: Pillai test statistic
#>               Df test stat approx F num Df den Df    Pr(>F)    
#> (Intercept)    1  0.220968  11.0621      2     78 5.897e-05 ***
#> Trend          1  0.031657   1.2750      2     78    0.2852    
#> Level Phase B  1  0.221748  11.1123      2     78 5.672e-05 ***
#> Slope Phase B  1  0.031469   1.2672      2     78    0.2874    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1