Overlap indices for single-case data

The overlapSC function provides the most common overlap indices for single-case data and some additional statistics.

overlap(data, dvar, pvar, mvar, decreasing = FALSE, phases = c(1, 2))

overlapSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

decreasing
If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

...
Further arguments passed to the function.

Value

overlap

A data frame consisting of the following indices for each single-case for all cases: PND, PEM, PET, NAP, PAND, Tau-U (A vs. B - Trend A), Diff_mean, Diff_trend, SMD.

phases.A

Selection for A phase.

phases.B

Selection for B phase.

design

Phase design.

See also

Examples

## Display overlap indices for one single-case
overlapSC(Huitema2000, decreasing = TRUE)

#> Overlap Indices
#> 
#> Design:  A B 
#> Comparing phase 1 against phase 2 
#> 
#>               Case1
#> PND           40.00
#> PEM           70.00
#> PET          100.00
#> NAP           70.00
#> NAP.rescaled  40.00
#> PAND          60.00
#> TAU_U         -0.68
#> Base_Tau      -0.73
#> Diff_mean    -20.40
#> Diff_trend   -17.87
#> SMD           -0.90
#> 
#> Note. The following variables were used in this analysis:
#>       'score' as independent variable, 'phase' as phase ,and 'mt' as measurement time.

## Display overlap indices for six single-cases
overlapSC(GruenkeWilbert2014)

#> Overlap Indices
#> 
#> Design:  A B 
#> Comparing phase 1 against phase 2 
#> 
#>               Anton    Bob   Paul Robert    Sam    Tim
#> PND          100.00 100.00 100.00 100.00 100.00 100.00
#> PEM          100.00 100.00 100.00 100.00 100.00 100.00
#> PET          100.00 100.00 100.00 100.00 100.00 100.00
#> NAP          100.00 100.00 100.00 100.00 100.00 100.00
#> NAP.rescaled 100.00 100.00 100.00 100.00 100.00 100.00
#> PAND         100.00 100.00 100.00 100.00 100.00 100.00
#> TAU_U          0.43   0.53   0.54   0.46   0.46   0.32
#> Base_Tau       0.67   0.77   0.75   0.77   0.72   0.66
#> Diff_mean      4.14   5.82   5.00   4.78   4.48   6.00
#> Diff_trend     0.43   0.00   0.28  -0.08  -0.07   0.60
#> SMD            5.07   7.13   6.64   5.72   8.17   7.35
#> 
#> Note. The following variables were used in this analysis:
#>       'score' as independent variable, 'phase' as phase ,and 'mt' as measurement time.

## Combining phases for analyszing designs with more than two phases   
overlapSC(exampleA1B1A2B2, phases = list(c("A1","A2"), c("B1","B2")))

#> Overlap Indices
#> 
#> Design:  A1 B1 A2 B2 
#> Comparing phases 1 + 3 against phases 2 + 4 
#> 
#>               Pawel Moritz Jannis
#> PND           55.00  77.78  71.43
#> PEM          100.00 100.00 100.00
#> PET          100.00 100.00 100.00
#> NAP           94.50  96.84  98.35
#> NAP.rescaled  89.00  93.69  96.70
#> PAND          82.50  85.00  90.00
#> TAU_U          0.44   0.45   0.37
#> Base_Tau       0.65   0.67   0.68
#> Diff_mean     12.25  13.58  15.27
#> Diff_trend    -0.05   0.00  -0.54
#> SMD            2.68   3.27   3.62