Percentage of all non-overlapping data

The pand function calculates the percentage of all non-overlapping data (PAND; Parker, Hagan-Burke, & Vannest, 2007), an index to quantify a level increase (or decrease) in performance after the onset of an intervention.

pand(data, dvar, pvar, decreasing = FALSE, correction = TRUE, phases = c(1, 2))

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

decreasing
If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

correction
The default correction = TRUE makes pand use a frequency matrix, which is corrected for ties. A tie is counted as the half of a measurement in both phases. Set correction = FALSE to use the uncorrected matrix, which is not recommended.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

Value

pand

Percentage of all non-overlapping data.

phi

Effect size Phi based on expected and observed values.

perc_overlap

Percentage of overlapping data points.

overlaps

Number of overlapping data points.

n

Number of data points.

N

Number of cases.

nA

Number of data points in phase A.

nB

Number of data points in phase B.

pA

Percentage of data points in phase A.

pB

Percentage of data points in phase B.

matrix

2x2 frequency matrix of phase A and B comparisons.

matrix_counts

2x2 counts matrix of phase A and B comparisons.

correlation

A list of the correlation values: statistic, parameter, p.value, estimate, null.value, alternative, method, data.name, correction.

correction

Logical argument from function call (see Arguments above).

Details

The PAND indicates nonoverlap between phase A and B data (like PND), but uses all data and is therefore not based on one single (probably unrepresentative) datapoint. Furthermore, PAND allows the comparison of real and expected associations (Chi-square test) and estimation of the effect size Phi, which equals Pearsons r for dichotomous data. Thus, phi-Square is the amount of explained variance. The original procedure for computing the PAND (Parker, Hagan-Burke, & Vannest, 2007) does not account for ambivalent datapoints (ties). The newer NAP overcomes this problem and has better precision-power (Parker, Vannest, & Davis, 2014).

References

Parker, R. I., Hagan-Burke, S., & Vannest, K. (2007). Percentage of All Non-Overlapping Data (PAND): An Alternative to PND. The Journal of Special Education, 40, 194-204.

Parker, R. I., & Vannest, K. (2009). An Improved Effect Size for Single-Case Research: Nonoverlap of All Pairs. Behavior Therapy, 40, 357-367.

See also

Author

Juergen Wilbert

Examples

## Calculate the PAND for a MMBD over three cases
gunnar <- scdf(c(2,3,1,5,3,4,2,6,4,7), B.start = 5)
birgit <- scdf(c(3,3,2,4,7,4,2,1,4,7), B.start = 4)
bodo   <- scdf(c(2,3,4,5,3,4,7,6,8,7), B.start = 6)
mbd <- c(gunnar, birgit, bodo)
pand(mbd)

#> Percentage of all non-overlapping data
#> 
#> PAND =  70 %
#> Φ =  0.375  ; Φ² =  0.141 
#> 
#> Number of cases: 3 
#> Total measurements: 30  (in phase A: 12; in phase B: 18)
#> n overlapping data per case: 3, 4, 2
#> Total overlapping data: n = 9 ; percentage = 30 
#> 
#> 2 x 2 Matrix of proportions
#> 	% expected
#> 	A	B	total
#> %    A	25	15	40
#> real B	15	45	60
#>  total	40	60
#> 
#> 2 x 2 Matrix of counts
#> 	expected
#> 	A	B	total
#>      A	7.5	4.5	12
#> real B	4.5	13.5	18
#>  total	12	18
#> 
#> 
#> Note. Matrix is corrected for ties
#> 
#> Correlation based analysis:
#> 
#> z = 2.393, p = 0.017, τ = 0.444 

pand(bodo)

#> Percentage of all non-overlapping data
#> 
#> PAND =  80 %
#> Φ =  0.6  ; Φ² =  0.36 
#> 
#> Number of cases: 1 
#> Total measurements: 10  (in phase A: 5; in phase B: 5)
#> n overlapping data per case: 2
#> Total overlapping data: n = 2 ; percentage = 20 
#> 
#> 2 x 2 Matrix of proportions
#> 	% expected
#> 	A	B	total
#> %    A	40	10	50
#> real B	10	40	50
#>  total	50	50
#> 
#> 2 x 2 Matrix of counts
#> 	expected
#> 	A	B	total
#>      A	4	1	5
#> real B	1	4	5
#>  total	5	5
#> 
#> 
#> Note. Matrix is corrected for ties
#> 
#> Correlation based analysis:
#> 
#> z = 1.800, p = 0.072, τ = 0.600 

## Calculate the PAND with an expected decrease of phase B scores
cubs <- scdf(c(20,22,24,17,21,13,10,9,20,9,18), B.start = 5)
pand(cubs, decreasing = TRUE)

#> Percentage of all non-overlapping data
#> 
#> PAND =  81.8 %
#> Φ =  0.607  ; Φ² =  0.369 
#> 
#> Number of cases: 1 
#> Total measurements: 11  (in phase A: 4; in phase B: 7)
#> n overlapping data per case: 2
#> Total overlapping data: n = 2 ; percentage = 18.2 
#> 
#> 2 x 2 Matrix of proportions
#> 	% expected
#> 	A	B	total
#> %    A	27.3	9.1	36.4
#> real B	9.1	54.5	63.6
#>  total	36.4	63.6
#> 
#> 2 x 2 Matrix of counts
#> 	expected
#> 	A	B	total
#>      A	3	1	4
#> real B	1	6	7
#>  total	4	7
#> 
#> 
#> Note. Matrix is corrected for ties
#> 
#> Correlation based analysis:
#> 
#> z = -1.807, p = 0.071, τ = -0.571