Percent exceeding the trend

The pet function returns the percentage of Phase B data points that exceed the prediction based on the Phase A trend. A binomial test against a 50/50 distribution is calculated. It also calculates the percentage of Phase B data points that exceed the upper (or lower) 95 percent confidence interval of the predicted progression.

Usage

pet(data, dvar, pvar, mvar, ci = 0.95, decreasing = FALSE, phases = c(1, 2))

Arguments

data

A single-case data frame. See scdf() to learn about this format.

dvar

Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar

Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar

Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

ci

Width of the confidence interval. Default is ci = 0.95.

decreasing

If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

phases

A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c(1,2).

Value

PET

Percent exceeding the trend.

PET.ci

Percent exceeding the upper / lower 95\ Test.

ci.percent

Width of confidence interval in percent.

se.factors

Standard error.

N

Number of cases.

decreasing

Logical argument from function call (see Arguments above).

case.names

Assigned name of single-case.

phases

-

See also

Author

Juergen Wilbert

Examples

## Calculate the PET and use a 99%-CI for the additional calculation
# create random example data
design <- design(n = 5, slope = 0.2)
dat <- random_scdf(design, seed = 23)
pet(dat, ci = .99)
#> Percent Exceeding the Trend
#> 
#> N cases =  5 
#> 
#>           PET binom.p PET CI
#> Case1  13.333   1.000  0.000
#> Case2 100.000   0.000 53.333
#> Case3 100.000   0.000 93.333
#> Case4   0.000   1.000  0.000
#> Case5  80.000   0.018  0.000
#> 
#> Binom.test: alternative hypothesis: true probability > 50%
#> PET CI: Percent of values greater than upper 99% confidence threshold (greater 2.326*se above predicted value)