Piecewise linear model / piecewise regression

The plm function computes a piecewise regression model (see Huitema & McKean, 2000).

plm(
  data,
  dvar,
  pvar,
  mvar,
  AR = 0,
  model = "B&L-B",
  family = "gaussian",
  trend = TRUE,
  level = TRUE,
  slope = TRUE,
  formula = NULL,
  update = NULL,
  na.action = na.omit,
  ...
)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

AR
Maximal lag of autoregression. Modeled based on the Autoregressive-Moving Average (ARMA) function. When AR is set, the family argument must be set to family = "gaussian".

model
Model used for calculating the slope parameter (see Huitema & McKean, 2000). Default is model = "B&L-B". Possible values are: "B&L-B", "H-M", "Mohr#1", "Mohr#2", "JW", "JW2", and "Manly".

family
Set the distributioin family. Defaults to a gaussian distribution. See the family function for more details.

trend
A logical indicating if a trend parameters is included in the model.

level
A logical indicating if a level parameters is included in the model.

slope
A logical indicating if a slope parameters is included in the model.

formula
Defaults to the standard piecewise regression model. The parameter phase followed by the phase name (e.g., phaseB) indicates the level effect of the corresponding phase. The parameter 'inter' followed by the phase name (e.g., interB) adresses the slope effect based on the method provide in the model argument (e.g., "B&L-B"). The formula can be changed for example to include further variables into the regression model.

update
An easier way to change the regression formula (e.g., . ~ . + newvariable).

na.action
Defines how to deal with missing values

...
Further arguments passed to the glm function.

Value

formula

plm formula. Uselful if you want to use the update or formula argument and you don't know the names of the parameters.

model

Character string from function call (see Arguments above).

F.test

F-test values of modelfit.

r.squares

Explained variance R squared for each model parameter.

ar

Autoregression lag from function call (see Arguments above).

family

Distribution family from function call (see Arguments above).

full.model

Full regression model list from the gls or glm function.

References

Beretvas, S., & Chung, H. (2008). An evaluation of modified R2-change effect size indices for single-subject experimental designs. Evidence-Based Communication Assessment and Intervention, 2, 120-128.

Huitema, B. E., & McKean, J. W. (2000). Design specification issues in time-series intervention models. Educational and Psychological Measurement, 60, 38-58.

See also

Author

Juergen Wilbert

Examples

## Compute a piecewise regression model for a random single-case
set.seed(123)
AB <- design_rSC(
  phase.design = list(A = 10, B = 20), 
  level = list(A = 0, B = 1), slope = list(A = 0, B = 0.05), 
  trend = list(0.05)
)
dat <- rSC(design = AB)
plm(dat, AR = 3)

#> Piecewise Regression Analysis
#> 
#> Dummy model:  B&L-B 
#> 
#> Fitted a gaussian distribution.
#> Correlated residuals up to autoregressions of lag 3 are modelled
#> 
#> F(3, 26) = 47.38; p = 0.000; R² = 0.845; Adjusted R² = 0.828
#> 
#>                    B   2.5%  97.5%    SE      t     p delta R²
#> Intercept     53.234 47.630 58.838 2.859 18.618 0.000         
#> Trend mt      -0.166 -1.032  0.701 0.442 -0.374 0.711  -0.0012
#> Level phase B 15.353  9.550 21.156 2.961  5.185 0.000   0.0735
#> Slope phase B  0.913 -0.045  1.871 0.489  1.869 0.073   0.0162
#> 
#> Autocorrelations of the residuals
#>  lag    cr
#>    1 -0.12
#>    2 -0.23
#>    3  0.38
#> 
#> Formula: values ~ 1 + mt + phaseB + interB
#> 

## Another example with a more complex design
A1B1A2B2 <- design_rSC(
  phase.design = list(A1 = 15, B1 = 20, A2 = 15, B2 = 20), 
  level = list(A1 = 0, B1 = 1, A2 = -1, B2 = 1),
  slope = list(A1 = 0, B1 = 0.0, A1 = 0, B2 = 0.0),
  trend = list(0.0))
dat <- rSC(design = A1B1A2B2, seed = 123)
plm(dat, model = "JW")

#> Piecewise Regression Analysis
#> 
#> Dummy model:  JW 
#> 
#> Fitted a gaussian distribution.
#> F(7, 62) = 11.05; p = 0.000; R² = 0.555; Adjusted R² = 0.505
#> 
#>                      B    2.5%  97.5%    SE      t     p delta R²
#> Intercept       51.439  46.427 56.451 2.557 20.115 0.000         
#> Trend mt        -0.085  -0.636  0.467 0.281 -0.301 0.764   0.0006
#> Level phase B1   7.536   1.297 13.775 3.183  2.367 0.021   0.0402
#> Level phase A2 -11.120 -17.517 -4.722 3.264 -3.407 0.001   0.0833
#> Level phase B2  10.261   4.022 16.500 3.183  3.224 0.002   0.0746
#> Slope phase B1   0.280  -0.377  0.937 0.335  0.835 0.407   0.0050
#> Slope phase A2  -0.240  -0.897  0.417 0.335 -0.716 0.477   0.0037
#> Slope phase B2   0.119  -0.538  0.776 0.335  0.354 0.724   0.0009
#> 
#> Autocorrelations of the residuals
#>  lag    cr
#>    1 -0.05
#>    2 -0.12
#>    3  0.12
#> 
#> Formula: values ~ 1 + mt + phaseB1 + phaseA2 + phaseB2 + interB1 + interA2 + 
#>     interB2
#> 

## no slope effects were found. Therefore you might want to drop slope estimation:
plm(dat, slope = FALSE, model = "JW")

#> Piecewise Regression Analysis
#> 
#> Dummy model:  JW 
#> 
#> Fitted a gaussian distribution.
#> F(4, 65) = 19.73; p = 0.000; R² = 0.548; Adjusted R² = 0.521
#> 
#>                      B    2.5%  97.5%    SE      t     p delta R²
#> Intercept       50.157  47.279 53.035 1.468 34.157 0.000         
#> Trend mt         0.076  -0.133  0.284 0.107  0.710 0.480   0.0035
#> Level phase B1   7.671   2.879 12.463 2.445  3.137 0.003   0.0684
#> Level phase A2 -10.945 -15.737 -6.153 2.445 -4.477 0.000   0.1392
#> Level phase B2   9.402   4.610 14.194 2.445  3.846 0.000   0.1028
#> 
#> Autocorrelations of the residuals
#>  lag    cr
#>    1 -0.03
#>    2 -0.09
#>    3  0.13
#> 
#> Formula: values ~ 1 + mt + phaseB1 + phaseA2 + phaseB2
#> 

## and now drop the trend estimation as well
plm(dat, slope = FALSE, trend = FALSE, model = "JW")

#> Piecewise Regression Analysis
#> 
#> Dummy model:  JW 
#> 
#> Fitted a gaussian distribution.
#> F(3, 66) = 26.34; p = 0.000; R² = 0.545; Adjusted R² = 0.524
#> 
#>                     B    2.5%  97.5%    SE      t p delta R²
#> Intercept      50.762  48.427 53.097 1.191 42.611 0         
#> Level phase B1  8.995   5.906 12.084 1.576  5.708 0   0.2247
#> Level phase A2 -9.621 -12.710 -6.532 1.576 -6.105 0   0.2570
#> Level phase B2 10.726   7.637 13.815 1.576  6.806 0   0.3195
#> 
#> Autocorrelations of the residuals
#>  lag    cr
#>    1 -0.01
#>    2 -0.08
#>    3  0.13
#> 
#> Formula: values ~ 1 + phaseB1 + phaseA2 + phaseB2
#>