Percentage of non-overlapping data

This function returns the percentage of non-overlapping data. Due to its error-proneness the PND should not be used, but nap or pand instead (see Parker & Vannest, 2009).

Usage

pnd(data, dvar, pvar, decreasing = FALSE, phases = c("A", "B"))

Arguments

data

A single-case data frame. See scdf to learn about this format.

dvar

Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar

Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

decreasing

If you expect data to be lower in the B phase, set decreasing = TRUE. Default is decreasing = FALSE.

phases

A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c(1,2).

Value

PND

Percentage of non-overlapping data.

See also

Author

Juergen Wilbert

Examples

## Calculate the PND for multiple single-case data
pnd(GruenkeWilbert2014)
#> Percent Non-Overlapping Data
#> 
#>    Case  PND Total Exceeds
#>   Anton 100%    14      14
#>     Bob 100%    11      11
#>    Paul 100%    12      12
#>  Robert 100%    10      10
#>     Sam 100%    13      13
#>     Tim 100%    14      14
#> 
#> Mean  : 100 %