Standardize values of an scdf file

This function scales the measured values of an scdf file. It allows for mean centering and standardization based on each single-case data set or a scaling across all cases included in an scdf.

standardize(
  data,
  var,
  center = TRUE,
  scale = FALSE,
  m = 0,
  sd = 1,
  grand = TRUE
)

Arguments

data

A single-case data frame. See scdf to learn about this format.

var

A character string or a vector of character strings with variable names that should be scaled.

center

If set TRUE, data are mean centered.

scale

If set TRUE, the standard deviation is set.

m

The target mean for centering.

sd

The target standard deviation for scaling

grand

If set TRUE, scaling is based on the mean and standard deviation of all values across all single-cases within the scdf.

Value

An scdf with the scaled values.

See also

Author

Juergen Wilbert

Examples

## Standardize a multiple case scdf and compute an hplm
ex_sc <- standardize(exampleAB_50, var = "values", center = TRUE, scale = TRUE)
hplm(ex_sc)
#> Hierarchical Piecewise Linear Regression
#> 
#> Estimation method ML 
#> Slope estimation method: B&L-B 
#> 50 Cases
#> 
#> ICC = 0.195; L = 192.0; p = 0.000
#> 
#> Fixed effects (values ~ 1 + mt + phaseB + interB)
#> 
#>                    B    SE   df       t p
#> Intercept     -1.346 0.068 1281 -19.775 0
#> Trend mt       0.036 0.006 1281   5.844 0
#> Level phase B  0.687 0.031 1281  21.942 0
#> Slope phase B  0.046 0.006 1281   7.236 0
#> 
#> Random effects (~1 | case)
#> 
#>           EstimateSD
#> Intercept      0.434
#> Residual       0.283