Tau-U for single-case data

This function calculates indices of the Tau-U family as proposed by Parker et al. (2011).

tau_u(
  data,
  dvar,
  pvar,
  tau_method = "b",
  method = "complete",
  phases = c(1, 2),
  fisher = TRUE,
  continuity_correction = FALSE
)

tauUSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

tau_method
Character with values "a" or "b" (default) indicating whether Kendall Tau A or Kendall Tau B is applied.

method
"complete" (default) or "parker". The latter calculates the number of possible pairs as described in Parker et al. (2011) which might lead to tau-U values greater than 1.

phases
A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

fisher
If TRUE, meta analyses of Tau U is based on Fisher Z transformed correlations.

continuity_correction
If TRUE, a continuity correction is applied for calculating p-values of correlations. This parameter is not yet implemented.

...
Further arguments passed to the function.

Value

table

A data frame containing statistics from the Tau-U family, including: Pairs, positive and negative comparisons, S, and Tau

matrix

The matrix of comparisons used for calculating the statistics.

tau_u

Tau-U value.

Details

Tau-U is an inconsistently operationalized construct.

References

Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42, 284-299.

See also

Author

Juergen Wilbert

Examples

## Calculate tau-U for the example from Parker et al. (2011)
bob <- scdf(c(A = 2, 3, 5, 3, B = 4, 5, 5, 7, 6), name = "Bob")
tau_u(bob)

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#>                        Model Tau_U    se    z df        p
#>                      A vs. B 0.842 0.103 8.19  0 2.63e-16
#>            A vs. B - Trend A 0.451 0.282 1.60  0 1.10e-01
#>            A vs. B + Trend B 0.742 0.159 4.68  0 2.92e-06
#>  A vs. B + Trend B - Trend A 0.589 0.231 2.55  0 1.07e-02
#> 
#> $Bob
#>                               Tau SE_Tau    Z     p
#> A vs. B                     0.842  0.430 1.96 0.050
#> A vs. B - Trend A           0.451  0.294 1.53 0.125
#> A vs. B + Trend B           0.742  0.288 2.58 0.010
#> A vs. B + Trend B - Trend A 0.589  0.275 2.14 0.032
#> 

## Calculate tau-U based on Kendall Tau A
tau_u(Grosche2011$Eva, tau_method = "a")

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#>                        Model Tau_U    se    z df       p
#>                      A vs. B 0.385 0.201 1.92  0 0.05548
#>            A vs. B - Trend A 0.355 0.206 1.72  0 0.08501
#>            A vs. B + Trend B 0.513 0.174 2.95  0 0.00316
#>  A vs. B + Trend B - Trend A 0.485 0.180 2.69  0 0.00706
#> 
#> $Eva
#>                               Tau SE_Tau    Z     p
#> A vs. B                     0.385  0.167 2.30 0.021
#> A vs. B - Trend A           0.355  0.167 2.12 0.034
#> A vs. B + Trend B           0.513  0.167 3.07 0.002
#> A vs. B + Trend B - Trend A 0.485  0.167 2.90 0.004
#> 

## Request tau-U for all single-cases fom the Grosche2011 data
tau_u(Grosche2011)

#> Tau-U
#> Method:  complete 
#> 
#> Overall Tau-U: 
#>                        Model  Tau_U    se     z df      p
#>                      A vs. B 0.0972 0.122 0.798  2 0.4251
#>            A vs. B - Trend A 0.0894 0.126 0.710  2 0.4777
#>            A vs. B + Trend B 0.2264 0.116 1.952  2 0.0509
#>  A vs. B + Trend B - Trend A 0.2447 0.116 2.117  2 0.0343
#> 
#> $Eva
#>                               Tau SE_Tau    Z     p
#> A vs. B                     0.385  0.292 1.32 0.188
#> A vs. B - Trend A           0.262  0.186 1.41 0.159
#> A vs. B + Trend B           0.490  0.172 2.85 0.004
#> A vs. B + Trend B - Trend A 0.485  0.167 2.90 0.004
#> 
#> $Georg
#>                                Tau SE_Tau      Z     p
#> A vs. B                     -0.042  0.265 -0.159 0.874
#> A vs. B - Trend A            0.056  0.164  0.342 0.732
#> A vs. B + Trend B            0.140  0.150  0.931 0.352
#> A vs. B + Trend B - Trend A  0.192  0.146  1.315 0.188
#> 
#> $Olaf
#>                                Tau SE_Tau      Z     p
#> A vs. B                     -0.104  0.270 -0.386 0.700
#> A vs. B - Trend A           -0.057  0.170 -0.336 0.737
#> A vs. B + Trend B           -0.091  0.177 -0.516 0.606
#> A vs. B + Trend B - Trend A -0.074  0.162 -0.454 0.650
#>