Tau-U for single-case data

This function calculates indices of the Tau-U family as proposed by Parker et al. (2011).

tau_u(
  data,
  dvar,
  pvar,
  tau_method = "b",
  method = "complete",
  phases = c(1, 2),
  meta_method = "random",
  ci = 0.95,
  continuity_correction = FALSE
)

Arguments

data

A single-case data frame. See scdf to learn about this format.

dvar

Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar

Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

tau_method

Character with values "a" or "b" (default) indicating whether Kendall Tau A or Kendall Tau B is applied.

method

"complete" (default) or "parker". The latter calculates the number of possible pairs as described in Parker et al. (2011) which might lead to tau-U values greater than 1.

phases

A vector of two characters or numbers indicating the two phases that should be compared. E.g., phases = c("A","C") or phases = c(2,4) for comparing the second to the fourth phase. Phases could be combined by providing a list with two elements. E.g., phases = list(A = c(1,3), B = c(2,4)) will compare phases 1 and 3 (as A) against 2 and 4 (as B). Default is phases = c("A","B").

meta_method

Character string. If set "random", a random-effect meta-analysis is calculated. If set "fixed", a fixed-effect meta-analysis is calculated. If set "none", no meta-analysis is conducted (may be helpful to speed up analyses).

ci

Confidence interval for meta analyzes.

continuity_correction

If TRUE, a continuity correction is applied for calculating p-values of correlations (here: S will be reduced by one before calculating Z)

Value

table

A data frame containing statistics from the Tau-U family, including: Pairs, positive and negative comparisons, S, and Tau

matrix

The matrix of comparisons used for calculating the statistics.

tau_u

Tau-U value.

Details

Tau-U is an inconsistently operationalized construct. Parker et al. (2011) describe a method which may result in Tau-U lager than 1. A different implementation of the method (provided at http://www.singlecaseresearch.org/calculators/tau-u) uses tau-b (instead of tau-a as in the original formulation by Parker). Bossart et. al (2018) describe inconsistencies in the results from this implementation as well. Another problems lies in the calculation in overall Tau-U values from several single cases. The function presented here applies a metaanalyzes to gain the overall values. Each tau value is weighted by the inverse of the variance (ie. the tau standard error). Tau values are not converted to Pearson r values. The argument "meta_method" calculates a random-effect model ("random") or a fixed effect model ("fixed").

References

Brossart, D. F., Laird, V. C., & Armstrong, T. W. (2018). Interpreting Kendall’s Tau and Tau-U for single-case experimental designs. Cogent Psychology, 5(1), 1–26. https://doi.org/10.1080/23311908.2018.1518687.

Parker, R. I., Vannest, K. J., Davis, J. L., & Sauber, S. B. (2011). Combining Nonoverlap and Trend for Single-Case Research: Tau-U. Behavior Therapy, 42, 284-299.

See also

Author

Juergen Wilbert

Examples

tau_u(Grosche2011$Eva)
#> Tau-U
#> Method: complete 
#> Applied Kendall's Tau-b
#> 
#> Case: Eva 
#>                               Tau SE_Tau    Z     p
#> A vs. B                     0.385  0.292 1.32 0.188
#> A vs. B - Trend A           0.262  0.186 1.41 0.159
#> A vs. B + Trend B           0.490  0.172 2.85 0.004
#> A vs. B + Trend B - Trend A 0.485  0.167 2.90 0.004
#> 

## Replicate  tau-U calculation from Parker et al. (2011)
bob <- scdf(c(A = 2, 3, 5, 3, B = 4, 5, 5, 7, 6), name = "Bob")
res <- tau_u(bob, method = "parker", tau_method = "a")
print(res, complete = TRUE)
#> Tau-U
#> Method: parker 
#> Applied Kendall's Tau-a
#> 95% CIs for tau are reported.
#> 
#> Case: Bob 
#>                             pairs pos neg ties  S  D   Tau  CI lower CI upper
#> A vs. B                        20  17   1    2 16 20 0.800 -0.000152     1.60
#> Trend A                         6   4   1    1  3  6 0.500 -0.461663     1.46
#> Trend B                        10   8   1    1  7 10 0.700 -0.100152     1.50
#> A vs. B - Trend A              20  18   5    3 13 20 0.650 -0.289966     1.59
#> A vs. B + Trend B              30  25   2    3 23 30 0.767  0.140023     1.39
#> A vs. B + Trend B - Trend A    36  26   6    4 20 36 0.556  0.033352     1.08
#>                             SD_S VAR_S SE_Tau    Z     p
#> A vs. B                     8.16 66.67  0.408 1.96 0.050
#> Trend A                     2.94  8.67  0.491 1.02 0.308
#> Trend B                     4.08 16.67  0.408 1.71 0.086
#> A vs. B - Trend A           9.59 92.00  0.480 1.36 0.175
#> A vs. B + Trend B           9.59 92.00  0.320 2.40 0.016
#> A vs. B + Trend B - Trend A 9.59 92.00  0.266 2.09 0.037
#> 

## Request tau-U for all single-cases from the Grosche2011 data set
tau_u(Grosche2011)
#> Tau-U
#> Method: complete 
#> Applied Kendall's Tau-b
#> 95% CIs for tau are reported.
#> 
#> Overall Tau-U
#> Meta-anlysis model: random effect
#> 
#>                        Model  Tau_U     se CI lower CI upper     z     p
#>                      A vs. B 0.0623 0.1587   -0.249    0.373 0.392 0.695
#>            A vs. B - Trend A 0.0762 0.0995   -0.119    0.271 0.765 0.444
#>            A vs. B + Trend B 0.1800 0.1637   -0.141    0.501 1.100 0.271
#>  A vs. B + Trend B - Trend A 0.1992 0.1572   -0.109    0.507 1.267 0.205
#> 
#> Case: Eva 
#>                               Tau SE_Tau    Z     p
#> A vs. B                     0.385  0.292 1.32 0.188
#> A vs. B - Trend A           0.262  0.186 1.41 0.159
#> A vs. B + Trend B           0.490  0.172 2.85 0.004
#> A vs. B + Trend B - Trend A 0.485  0.167 2.90 0.004
#> 
#> Case: Georg 
#>                                Tau SE_Tau      Z     p
#> A vs. B                     -0.042  0.265 -0.159 0.874
#> A vs. B - Trend A            0.056  0.164  0.342 0.732
#> A vs. B + Trend B            0.140  0.150  0.931 0.352
#> A vs. B + Trend B - Trend A  0.192  0.146  1.315 0.188
#> 
#> Case: Olaf 
#>                                Tau SE_Tau      Z     p
#> A vs. B                     -0.104  0.270 -0.386 0.700
#> A vs. B - Trend A           -0.057  0.170 -0.336 0.737
#> A vs. B + Trend B           -0.091  0.177 -0.516 0.606
#> A vs. B + Trend B - Trend A -0.074  0.162 -0.454 0.650
#>