Transform variables in every single case of a single case data frame

Takes a scdf and applies transformations to each individual case. This is useful to calculate or modify new variables.

# S3 method for scdf
transform(`_data`, ...)

Arguments

_data

A scdf.

...

Expressions.

Value

A scdf.

Details

This function is a method of the generic transformation function. Unlike this one, expressions are calculated serially. This means that the results of the calculation of an expression are the basis for the following calculations.

Examples

## Creates a single-case with frequency distributions. The proportion and
## percentage of the frequencies are calculated with transform:
design <- design(
 n = 3,
 level = 5,
 distribution = "binomial",
 n_trials = 20,
 start_value = 0.5
)
study <- random_scdf(design)
transform(study, proportion = values/trials, percentage = proportion * 100)
#> #A single-case data frame with 3 cases
#> 
#>  Case1: phase values mt trials proportion percentage
#>             A     13  1     20       0.65         65
#>             A     14  2     20        0.7         70
#>             A      9  3     20       0.45         45
#>             A      7  4     20       0.35         35
#>             A     12  5     20        0.6         60
#>             B     20  6     20          1        100
#>             B     20  7     20          1        100
#>             B     20  8     20          1        100
#>             B     20  9     20          1        100
#>             B     20 10     20          1        100
#>             B     20 11     20          1        100
#>             B     20 12     20          1        100
#>             B     20 13     20          1        100
#>             B     20 14     20          1        100
#>             B     20 15     20          1        100
#> # ... up to 5 more rows
#> #  2 more cases

## Z standardize the dependent variable and add two new variables:
exampleAB %>% 
  transform(
    values = scale(values), 
    mean_values = mean(values), 
    sd_values = sd(values)
  )
#> #A single-case data frame with 3 cases
#> 
#>  Johanna: values mt phase mean_values sd_values
#>            -1.31  1     A           0         1
#>             -1.4  2     A           0         1
#>            -1.14  3     A           0         1
#>            -0.97  4     A           0         1
#>            -1.48  5     A           0         1
#>            -0.71  6     B           0         1
#>            -0.62  7     B           0         1
#>             0.15  8     B           0         1
#>            -0.28  9     B           0         1
#>            -0.45 10     B           0         1
#>             0.75 11     B           0         1
#>             0.06 12     B           0         1
#>             0.41 13     B           0         1
#>              1.1 14     B           0         1
#>             0.67 15     B           0         1
#> # ... up to 5 more rows
#> #  2 more cases