Trend analysis for single-cases data

The trend function provides an overview of linear trends in single-case data. By default, it gives you the intercept and slope of a linear and a squared regression of measurement-time on scores. Models are computed separately for each phase and across all phases. For a more advanced application, you can add regression models using the R specific formula class.

trend(data, dvar, pvar, mvar, offset = -1, model = NULL)

trendSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

offset
An offset for the first measurement-time of each phase (MT). If offset = 0, the phase measurement is handled as MT 1. Default is offset = -1, setting the first value of MT to 0.

model
A string or a list of (named) strings each depicting one regression model. This is a formula expression of the standard R class. The parameters of the model are values, mt and phase.

...
Further arguments passed to the function.

Value

trend

A matrix containing the results (Intercept, B and beta) of separate regression models for phase A, phase B, and the whole data.

offset

Numeric argument from function call (see Arguments section).

See also

Author

Juergen Wilbert

Examples

## Compute the linear and squared regression for a random single-case
design <- design_rSC(slope = 0.5)
matthea <- rSC(design)
trend(matthea)
#> Trend for each phase
#> 
#>             Intercept      B   Beta
#> Linear.ALL     41.211  4.200  0.931
#> Linear.A       63.724 -4.433 -0.880
#> Linear.B       51.745  5.445  0.978
#> Squared.ALL    53.534  0.223  0.974
#> Squared.A      61.055 -1.033 -0.855
#> Squared.B      64.756  0.371  0.967
#> 
#> Note. Measurement-times start at 0  for each phase

## Besides the linear and squared regression models compute two custom models:
## a) a cubic model, and b) the values predicted by the natural logarithm of the
## measurement time.
design <- design_rSC(slope = 0.3)
ben <- rSC(design)
trend(ben, offset = 0, model = c("Cubic" = values ~ I(mt^3), "Log Time" = values ~ log(mt)))
#> Trend for each phase
#> 
#>              Intercept      B   Beta
#> Linear.ALL      44.232  2.425  0.948
#> Linear.A        54.100 -0.466 -0.442
#> Linear.B        53.419  2.742  0.934
#> Squared.ALL     53.459  0.113  0.956
#> Squared.A       53.245 -0.049 -0.286
#> Squared.B       61.728  0.165  0.923
#> Cubic.ALL       57.309  0.006  0.930
#> Cubic.A         52.921 -0.005 -0.149
#> Cubic.B         64.987  0.011  0.897
#> Log Time.ALL    37.297 15.303  0.822
#> Log Time.A      54.154 -1.515 -0.579
#> Log Time.B      48.070 14.668  0.873
#> 
#> Note. Measurement-times start at 1  for each phase