Trend analysis for single-cases data

The trend function provides an overview of linear trends in single-case data. By default, it gives you the intercept and slope of a linear and a squared regression of measurement-time on scores. Models are computed separately for each phase and across all phases. For a more advanced application, you can add regression models using the R specific formula class.

trend(data, dvar, pvar, mvar, offset = -1, model = NULL)

Arguments

data

A single-case data frame. See scdf to learn about this format.

dvar

Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar

Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar

Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

offset

An offset for the first measurement-time of each phase (MT). If offset = 0, the phase measurement is handled as MT 1. Default is offset = -1, setting the first value of MT to 0.

model

A string or a list of (named) strings each depicting one regression model. This is a formula expression of the standard R class. The parameters of the model are values, mt and phase.

Value

trend

A matrix containing the results (Intercept, B and beta) of separate regression models for phase A, phase B, and the whole data.

offset

Numeric argument from function call (see Arguments section).

See also

Author

Juergen Wilbert

Examples

## Compute the linear and squared regression for a random single-case
design <- design(slope = 0.5)
matthea <- random_scdf(design)
trend(matthea)
#> Trend for each phase
#> 
#>             Intercept      B   Beta
#> Linear.ALL     39.896  4.320  0.970
#> Linear.A       53.985 -2.263 -0.932
#> Linear.B       58.390  4.720  0.978
#> Squared.ALL    54.124  0.217  0.960
#> Squared.A      52.674 -0.536 -0.921
#> Squared.B      70.613  0.308  0.926
#> 
#> Note. Measurement-times start at 0  for each phase

## Besides the linear and squared regression models compute two custom models:
## a) a cubic model, and b) the values predicted by the natural logarithm of the
## measurement time.
design <- design(slope = 0.3)
ben <- random_scdf(design)
trend(ben, offset = 0, model = c("Cubic" = values ~ I(mt^3), "Log Time" = values ~ log(mt)))
#> Trend for each phase
#> 
#>              Intercept      B  Beta
#> Linear.ALL      37.921  2.929 0.959
#> Linear.A        48.522  0.407 0.260
#> Linear.B        46.505  3.561 0.967
#> Squared.ALL     48.852  0.138 0.977
#> Squared.A       48.960  0.071 0.278
#> Squared.B       57.303  0.214 0.956
#> Cubic.ALL       53.453  0.007 0.957
#> Cubic.A         49.193  0.012 0.252
#> Cubic.B         61.605  0.014 0.924
#> Log Time.ALL    29.995 18.275 0.821
#> Log Time.A      49.093  0.678 0.175
#> Log Time.B      39.974 18.828 0.894
#> 
#> Note. Measurement-times start at 1  for each phase