Trend analysis for single-cases data

The trendSC function provides an overview of linear trends in single-case data. By default, it gives you the intercept and slope of a linear and a squared regression of measurement-time on scores. Models are computed separately for each phase and across all phases. For a more advanced application, you can add regression models using the R specific formula class.

trend(data, dvar, pvar, mvar, offset = -1, model = NULL)

trendSC(...)

Arguments

data
A single-case data frame. See scdf to learn about this format.

dvar
Character string with the name of the dependent variable. Defaults to the attributes in the scdf file.

pvar
Character string with the name of the phase variable. Defaults to the attributes in the scdf file.

mvar
Character string with the name of the measurement time variable. Defaults to the attributes in the scdf file.

offset
An offset for the first measurement-time of each phase (MT). If set offset = 0, the phase measurement is handled as MT 1. Default is offset = -1, setting the first value of MT to 0.

model
A string or a list of (named) strings each depicting one regression model. This is a formula expression of the standard R class. The parameters of the model are values, mt and phase.

...
Further arguments passed to the function.

Value

trend

A matrix containing the results (Intercept, B and beta) of separate regression models for phase A, phase B, and the whole data.

offset

Numeric argument from function call (see Arguments section).

See also

Author

Juergen Wilbert

Examples

## Compute the linear and squared regression for a random single-case
design <- design_rSC(slope = 0.5)
matthea <- rSC(design)
trendSC(matthea)

#> Trend for each phase
#> 
#>             Intercept      B   Beta
#> Linear.ALL     41.283  4.319  0.962
#> Linear.A       54.346 -1.183 -0.285
#> Linear.B       58.241  4.883  0.970
#> Squared.ALL    55.067  0.221  0.967
#> Squared.A      53.038 -0.176 -0.177
#> Squared.B      70.139  0.329  0.950
#> 
#> Note. Measurement-times of phase B start at 0 

## Besides the linear and squared regression models compute two custom models:
## a) a cubic model, and b) the values predicted by the natural logarithm of the
## measurement time.
design <- design_rSC(slope = 0.3)
ben <- rSC(design)
trend(ben, offset = 0, model = c("Cubic" = values ~ I(mt^3), "Log Time" = values ~ log(mt)))

#> Trend for each phase
#> 
#>              Intercept      B  Beta
#> Linear.ALL      37.217  2.908 0.974
#> Linear.A        43.401  1.416 0.719
#> Linear.B        48.324  3.266 0.966
#> Squared.ALL     48.573  0.134 0.967
#> Squared.A       45.534  0.192 0.597
#> Squared.B       58.568  0.192 0.935
#> Cubic.ALL       53.293  0.007 0.930
#> Cubic.A         46.283  0.030 0.497
#> Cubic.B         62.649  0.012 0.887
#> Log Time.ALL    27.606 18.966 0.872
#> Log Time.A      43.774  4.047 0.826
#> Log Time.B      41.479 17.729 0.917
#> 
#> Note. Measurement-times of phase B start at 1